pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["Moha-AlHanbali <moha.alhanbali@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
Django = "^3.2.9"

[tool.poetry.dev-dependencies]
Django = "^3.2.9"
djangorestframework = "^3.12.4"
flake8 = "^4.0.1"
black = "^21.11b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
